Fixed some bugs
Database would get locked sometimes, added `defer row.Close()`, it should close by itself without it when there are no more rows, but that doesn't always happen. Changed `getQuestion` to not send unneeded data in the answers. Made `endRun` respond with a `bool` corresponding to whether the query affected the run or not.